fix(#111): extract session entry types to a leaf module to break madge cycle
check-additional-runtime-topology-architecture (pnpm check:architecture ->
check:madge-import-cycles) flagged a 1-node-pair cycle introduced by r7's
extraction of strip-stale-thinking-blocks.ts:

  strip-stale-thinking-blocks.ts -> session-manager.ts

session-manager.ts imports stripStaleThinkingBlocksFromSessionManagerBranch
from strip-stale-thinking-blocks.ts at runtime (r7's #111 fix, intentional --
avoids a separate, worse cycle through transcript-rewrite.ts ->
transcript-file-state.ts -> session-manager.ts). strip-stale-thinking-blocks.ts
in turn imported { is a shell keyword
} is a shell keyword (for the SessionManagerLike
structural type) and, after this fix round's item A, { is a shell keyword from session-manager.ts. Madge's
cycle checker (scripts/check-madge-import-cycles.ts) does not distinguish
type-only import edges from runtime value edges, so this closed a cycle even
though the type-only edge carries no runtime dependency.

Fix (module placement, not a rules change, per the architecture check's own
intent): extracted the pure entry-type declarations (SessionHeader,
SessionEntryBase and all its variants, SessionEntry, FileEntry,
AppendPersistenceOptions, NewSessionOptions) into a new leaf module
session-entry-types.ts that depends on nothing from session-manager.ts or
anything it imports. session-manager.ts now imports these types from the new
leaf module and re-exports them (so every existing importer of these types
from session-manager.js is unaffected). strip-stale-thinking-blocks.ts now
imports SessionEntry/AppendPersistenceOptions from session-entry-types.ts
directly, and defines SessionManagerLike as an explicit local structural
interface (method signatures only) instead of deriving it via
 -- removing the last import of the
SessionManager class/value from this module entirely.

Verified: pnpm check:architecture passes in full (0 runtime cycles, 0 madge
cycles, all other architecture sub-checks green). pnpm tsgo:core,
tsgo:core:test, tsgo:extensions:test all clean. pnpm lint clean (0
errors across core/extensions/scripts shards). Full test suites for touched
areas pass: session-manager.test.ts (47), transcript-rewrite.test.ts (7),
compaction-successor-transcript.test.ts (13), context-engine-maintenance.test.ts
(21), tool-result-truncation.test.ts (38), thinking-replay-repair.test.ts (4).
