+++ src/agents/sessions/session-entry-types.ts
@@ -1,0 +1,164 @@
+/**
+ * Session entry type declarations (SessionEntry/FileEntry and friends).
+ *
+ * Extracted from session-manager.ts (#111 C5 architecture fix) so leaf
+ * modules that session-manager.ts imports at runtime -- e.g.
+ * embedded-agent-runner/strip-stale-thinking-blocks.ts, which
+ * session-manager.ts calls for path-independent stale-thinking-block cleanup
+ * -- can reference these types without importing session-manager.ts itself.
+ * Importing session-manager.ts from such a module, even via `import type`,
+ * closes a cycle: pnpm check:architecture's madge-based cycle check does not
+ * distinguish type-only import edges from runtime value edges, so any import
+ * of session-manager.ts from a module session-manager.ts imports at runtime
+ * is flagged as a circular dependency regardless of whether the edge is
+ * type-only.
+ *
+ * This module has no dependency on session-manager.ts (or anything it
+ * imports), so it is safe for both session-manager.ts and any of the
+ * modules it imports at runtime to depend on.
+ */
+import type { ImageContent, TextContent } from "../../llm/types.js";
+import type { AgentMessage } from "../runtime/index.js";
+
+export interface SessionHeader {
+  type: "session";
+  version?: number; // v1 sessions don't have this
+  id: string;
+  timestamp: string;
+  cwd: string;
+  parentSession?: string;
+}
+
+export interface NewSessionOptions {
+  id?: string;
+  parentSession?: string;
+}
+
+export interface SessionEntryBase {
+  type: string;
+  id: string;
+  parentId: string | null;
+  timestamp: string;
+  /** This row consumes the raw side cursor instead of the visible leaf. */
+  appendMode?: "side";
+}
+
+export interface SessionMessageEntry extends SessionEntryBase {
+  type: "message";
+  message: AgentMessage;
+}
+
+export interface ThinkingLevelChangeEntry extends SessionEntryBase {
+  type: "thinking_level_change";
+  thinkingLevel: string;
+}
+
+export interface ModelChangeEntry extends SessionEntryBase {
+  type: "model_change";
+  provider: string;
+  modelId: string;
+}
+
+export interface CompactionEntry<T = unknown> extends SessionEntryBase {
+  type: "compaction";
+  summary: string;
+  firstKeptEntryId: string;
+  tokensBefore: number;
+  /** Extension-specific data (e.g., ArtifactIndex, version markers for structured compaction) */
+  details?: T;
+  /** True if generated by an extension, undefined/false if runtime-generated (backward compatible) */
+  fromHook?: boolean;
+}
+
+export interface BranchSummaryEntry<T = unknown> extends SessionEntryBase {
+  type: "branch_summary";
+  fromId: string;
+  summary: string;
+  /** Extension-specific data (not sent to LLM) */
+  details?: T;
+  /** True if generated by an extension, false if runtime-generated */
+  fromHook?: boolean;
+}
+
+/**
+ * Custom entry for extensions to store extension-specific data in the session.
+ * Use customType to identify your extension's entries.
+ *
+ * Purpose: Persist extension state across session reloads. On reload, extensions can
+ * scan entries for their customType and reconstruct internal state.
+ *
+ * Does NOT participate in LLM context (ignored by buildSessionContext).
+ * For injecting content into context, see CustomMessageEntry.
+ */
+export interface CustomEntry<T = unknown> extends SessionEntryBase {
+  type: "custom";
+  customType: string;
+  data?: T;
+}
+
+/** Label entry for user-defined bookmarks/markers on entries. */
+export interface LabelEntry extends SessionEntryBase {
+  type: "label";
+  targetId: string;
+  label: string | undefined;
+}
+
+/** Session metadata entry (e.g., user-defined display name). */
+export interface SessionInfoEntry extends SessionEntryBase {
+  type: "session_info";
+  name?: string;
+}
+
+/**
+ * Custom message entry for extensions to inject messages into LLM context.
+ * Use customType to identify your extension's entries.
+ *
+ * Unlike CustomEntry, this DOES participate in LLM context.
+ * The content is converted to a user message in buildSessionContext().
+ * Use details for extension-specific metadata (not sent to LLM).
+ *
+ * display controls TUI rendering:
+ * - false: hidden entirely
+ * - true: rendered with distinct styling (different from user messages)
+ */
+export interface CustomMessageEntry<T = unknown> extends SessionEntryBase {
+  type: "custom_message";
+  customType: string;
+  content: string | (TextContent | ImageContent)[];
+  details?: T;
+  display: boolean;
+}
+
+/** Session entry - has id/parentId for tree structure (returned by "read" methods in SessionManager) */
+export type SessionEntry =
+  | SessionMessageEntry
+  | ThinkingLevelChangeEntry
+  | ModelChangeEntry
+  | CompactionEntry
+  | BranchSummaryEntry
+  | CustomEntry
+  | CustomMessageEntry
+  | LabelEntry
+  | SessionInfoEntry;
+
+/** Raw file entry (includes header) */
+export type FileEntry = SessionHeader | SessionEntry;
+
+/**
+ * Options accepted by SessionManager's appendXXX() persistence methods.
+ * Exported from this leaf module (not session-manager.ts) so callers that
+ * session-manager.ts imports at runtime can reference the exact options
+ * shape without importing session-manager.ts itself (see module doc above).
+ */
+export type AppendPersistenceOptions = {
+  invalidateSerializedPrefixCache?: boolean;
+  /**
+   * Explicit entry ID to use instead of generating a new one. Used by
+   * maintenance rewrite-replay paths (e.g.
+   * rewriteTranscriptEntriesInSessionManager) to replay an entry that was
+   * NOT the target of a replacement without reassigning it a new ID, so
+   * that callers holding a reference to that entry's original ID (e.g. a
+   * captured firstKeptEntryId) remain valid after the replay (#111).
+   */
+  entryId?: string;
+};
