+++ src/agents/sessions/session-manager.ts
@@ -49,100 +49,50 @@ import {
 } from "../runtime/index.js";
 import { invalidateSessionFileRepairCache } from "../session-file-repair.js";
 import type { BashExecutionMessage, CustomMessage } from "./messages.js";
+// Imported AND re-exported from session-entry-types.ts (#111 C5 architecture
+// fix) so existing importers of these types from session-manager.js keep
+// working unchanged. See session-entry-types.ts's module doc for why these
+// type declarations were extracted to their own leaf module.
+import type {
+  AppendPersistenceOptions,
+  BranchSummaryEntry,
+  CompactionEntry,
+  CustomEntry,
+  CustomMessageEntry,
+  FileEntry,
+  LabelEntry,
+  ModelChangeEntry,
+  NewSessionOptions,
+  SessionEntry,
+  SessionEntryBase,
+  SessionHeader,
+  SessionInfoEntry,
+  SessionMessageEntry,
+  ThinkingLevelChangeEntry,
+} from "./session-entry-types.js";
+
+export type {
+  AppendPersistenceOptions,
+  BranchSummaryEntry,
+  CompactionEntry,
+  CustomEntry,
+  CustomMessageEntry,
+  FileEntry,
+  LabelEntry,
+  ModelChangeEntry,
+  NewSessionOptions,
+  SessionEntry,
+  SessionEntryBase,
+  SessionHeader,
+  SessionInfoEntry,
+  SessionMessageEntry,
+  ThinkingLevelChangeEntry,
+};
 
 export { CURRENT_SESSION_VERSION };
 
 const SESSION_HEADER_READ_CHUNK_BYTES = 4096;
 const MAX_SESSION_HEADER_BYTES = 64 * 1024;
-
-export interface SessionHeader {
-  type: "session";
-  version?: number; // v1 sessions don't have this
-  id: string;
-  timestamp: string;
-  cwd: string;
-  parentSession?: string;
-}
-
-export interface NewSessionOptions {
-  id?: string;
-  parentSession?: string;
-}
-
-export interface SessionEntryBase {
-  type: string;
-  id: string;
-  parentId: string | null;
-  timestamp: string;
-  /** This row consumes the raw side cursor instead of the visible leaf. */
-  appendMode?: "side";
-}
-
-export interface SessionMessageEntry extends SessionEntryBase {
-  type: "message";
-  message: AgentMessage;
-}
-
-export interface ThinkingLevelChangeEntry extends SessionEntryBase {
-  type: "thinking_level_change";
-  thinkingLevel: string;
-}
-
-export interface ModelChangeEntry extends SessionEntryBase {
-  type: "model_change";
-  provider: string;
-  modelId: string;
-}
-
-export interface CompactionEntry<T = unknown> extends SessionEntryBase {
-  type: "compaction";
-  summary: string;
-  firstKeptEntryId: string;
-  tokensBefore: number;
-  /** Extension-specific data (e.g., ArtifactIndex, version markers for structured compaction) */
-  details?: T;
-  /** True if generated by an extension, undefined/false if runtime-generated (backward compatible) */
-  fromHook?: boolean;
-}
-
-export interface BranchSummaryEntry<T = unknown> extends SessionEntryBase {
-  type: "branch_summary";
-  fromId: string;
-  summary: string;
-  /** Extension-specific data (not sent to LLM) */
-  details?: T;
-  /** True if generated by an extension, false if runtime-generated */
-  fromHook?: boolean;
-}
-
-/**
- * Custom entry for extensions to store extension-specific data in the session.
- * Use customType to identify your extension's entries.
- *
- * Purpose: Persist extension state across session reloads. On reload, extensions can
- * scan entries for their customType and reconstruct internal state.
- *
- * Does NOT participate in LLM context (ignored by buildSessionContext).
- * For injecting content into context, see CustomMessageEntry.
- */
-export interface CustomEntry<T = unknown> extends SessionEntryBase {
-  type: "custom";
-  customType: string;
-  data?: T;
-}
-
-/** Label entry for user-defined bookmarks/markers on entries. */
-export interface LabelEntry extends SessionEntryBase {
-  type: "label";
-  targetId: string;
-  label: string | undefined;
-}
-
-/** Session metadata entry (e.g., user-defined display name). */
-export interface SessionInfoEntry extends SessionEntryBase {
-  type: "session_info";
-  name?: string;
-}
 
 interface PromptReleasedOpaqueEntry {
   type: "prompt_released_opaque";
@@ -160,54 +110,6 @@ type PromptReleasedSessionMergeResult = {
   sessionFileSnapshot: OwnedSessionTranscriptCacheSnapshot;
   publishedEntries?: readonly OwnedSessionTranscriptPublishedEntry[];
   requiresReload?: true;
-};
-
-/**
- * Custom message entry for extensions to inject messages into LLM context.
- * Use customType to identify your extension's entries.
- *
- * Unlike CustomEntry, this DOES participate in LLM context.
- * The content is converted to a user message in buildSessionContext().
- * Use details for extension-specific metadata (not sent to LLM).
- *
- * display controls TUI rendering:
- * - false: hidden entirely
- * - true: rendered with distinct styling (different from user messages)
- */
-export interface CustomMessageEntry<T = unknown> extends SessionEntryBase {
-  type: "custom_message";
-  customType: string;
-  content: string | (TextContent | ImageContent)[];
-  details?: T;
-  display: boolean;
-}
-
-/** Session entry - has id/parentId for tree structure (returned by "read" methods in SessionManager) */
-export type SessionEntry =
-  | SessionMessageEntry
-  | ThinkingLevelChangeEntry
-  | ModelChangeEntry
-  | CompactionEntry
-  | BranchSummaryEntry
-  | CustomEntry
-  | CustomMessageEntry
-  | LabelEntry
-  | SessionInfoEntry;
-
-/** Raw file entry (includes header) */
-export type FileEntry = SessionHeader | SessionEntry;
-
-type AppendPersistenceOptions = {
-  invalidateSerializedPrefixCache?: boolean;
-  /**
-   * Explicit entry ID to use instead of generating a new one. Used by
-   * maintenance rewrite-replay paths (e.g.
-   * rewriteTranscriptEntriesInSessionManager) to replay an entry that was
-   * NOT the target of a replacement without reassigning it a new ID, so
-   * that callers holding a reference to that entry's original ID (e.g. a
-   * captured firstKeptEntryId) remain valid after the replay (#111).
-   */
-  entryId?: string;
 };
 
 /** Tree node for getTree() - defensive copy of session structure */
